datapack model tweaks
